# catalog-service/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8083
    description: Generated server url
paths:
...
  /api/products/filter:
    get:
      x-spring-paginated: true
      tags:
        - product-rest-controller
      operationId: filterProducts
      parameters:
        - name: filterDto
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProductFilterDto"
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponseDto"
components:
  schemas:
...
    ProductResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
    ProductFilterDto:
      type: object
      properties:
        price:
          type: array
          items:
            type: number
            format: double
        name:
          type: string
        category:
          type: string
...